+++ src/types/plugins.d.ts
@@ -1,7 +1,6 @@
 declare module "virtual:papyrus-md-txt/config" {
   const config: {
     includeDrafts: boolean;
-    format: string;
   };
   export default config;
 }